bookstack-rag/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
sentence-transformers==5.4.1
qdrant-client==1.17.1
watchdog==6.0.0
python-frontmatter==1.1.0
